Conflicts:
	src/Oro/Bundle/BatchBundle/Tests/Unit/Entity/JobExecutionTest.php

+++ Bundle/WindowsBundle/Resources/public/js/dialog-widget.js
@@ -25,6 +25,7 @@ function(_, Backbone, app, error, AbstractWidget, StateModel) {
         windowY: 0,
         defaultPos: 'center center',
         openedWindows: 0,
+        contentTop: null,
 
         /**
          * Initialize dialog
@@ -39,7 +40,9 @@ function(_, Backbone, app, error, AbstractWidget, StateModel) {
             this.options.dialogOptions.title = this.options.dialogOptions.title || this.options.title;
             this.options.dialogOptions.limitTo = this.options.dialogOptions.limitTo || '#container';
 
-            this._initModel(this.options);
+            if (this.options.stateEnabled) {
+                this._initModel(this.options);
+            }
 
             var runner = function(handlers) {
                 return function() {
@@ -58,6 +61,7 @@ function(_, Backbone, app, error, AbstractWidget, StateModel) {
 
             this.options.dialogOptions.close = runner(closeHandlers);
 
+            this.on('widgetRender', _.bind(this._initAdjustHeight, this));
             this.on('contentLoadError', _.bind(this.loadErrorHandler, this));
         },
 
@@ -66,7 +70,7 @@ function(_, Backbone, app, error, AbstractWidget, StateModel) {
         },
 
         _initModel: function(options) {
-            if (this.options.stateEnabled && this.model) {
+            if (this.model) {
                 this.restoreMode = true;
                 var attributes = this.model.get('data');
                 _.extend(options, attributes);
@@ -91,14 +95,16 @@ function(_, Backbone, app, error, AbstractWidget, StateModel) {
          * Handle dialog close
          */
         closeHandler: function() {
-            this.model.destroy({
-                error: _.bind(function(model, xhr, options) {
-                    // Suppress error if it's 404 response and not debug mode
-                    if (xhr.status != 404 || app.debug) {
-                        error.dispatch(model, xhr, options);
-                    }
-                }, this)
-            });
+            if (this.model) {
+                this.model.destroy({
+                    error: _.bind(function(model, xhr, options) {
+                        // Suppress error if it's 404 response and not debug mode
+                        if (xhr.status != 404 || app.debug) {
+                            error.dispatch(model, xhr, options);
+                        }
+                    }, this)
+                });
+            }
             this.widget.remove();
             AbstractWidget.prototype.remove.call(this);
         },
@@ -126,7 +132,9 @@ function(_, Backbone, app, error, AbstractWidget, StateModel) {
             saveData.dialogOptions.state = data.state;
             saveData.dialogOptions.snapshot = data.snapshot;
 
-            this.model.save({data: saveData});
+            if (this.model) {
+                this.model.save({data: saveData});
+            }
         },
 
         remove: function() {
@@ -140,7 +148,9 @@ function(_, Backbone, app, error, AbstractWidget, StateModel) {
 
         loadErrorHandler: function()
         {
-            this.model.destroy();
+            if (this.model) {
+                this.model.destroy();
+            }
         },
 
         getActionsElement: function() {
@@ -177,6 +187,27 @@ function(_, Backbone, app, error, AbstractWidget, StateModel) {
                 this.widget.html(this.$el);
             }
             AbstractWidget.prototype.show.apply(this);
+        },
+
+        _initAdjustHeight: function(content) {
+            this.widget.off("dialogresize dialogmaximize dialogrestore", _.bind(this._fixScrollableHeight, this));
+            var scrollableContent = content.find('.scrollable-container');
+            if (scrollableContent.length) {
+                scrollableContent.css('overflow', 'auto');
+                this.widget.on("dialogresize dialogmaximize dialogrestore", _.bind(this._fixScrollableHeight, this));
+                this._fixScrollableHeight();
+            }
+        },
+
+        _fixScrollableHeight: function() {
+            var widget = this.widget;
+            widget.find('.scrollable-container').each(_.bind(function(i, el){
+                var $el = $(el);
+                var height = widget.height() - $el.position().top;
+                if (height) {
+                    $el.height(height);
+                }
+            },this));
         },
 
         /**
